ppt-translate/requirements.txt
python-pptx>=0.6.23
anthropic>=0.34.0
lxml>=5.0.0
flask>=3.0.0
msal>=1.28.0
requests>=2.31.0
slack_sdk>=3.27.0
